renamed "module classes" to "module_classes"
