+++ src/module_classes/Permission.js
