+++ src/module_classes/Snippet.js
